Extract shared node-red logs behaviour to a component
